+++ ambisonic_decode/type_definitions.py
@@ -1,2 +1,11 @@
+from typing import NamedTuple
+import numpy as np
+
 Rad = float  # angles in radian
+
+
+class TestData(NamedTuple):
+    input: np.ndarray
+    decoded: np.ndarray
+    order: int = 1
